historic-ocr/requirements.txt
# Podstawowe zależności
Flask==3.0.0
Flask-CORS==4.0.0
gunicorn==21.2.0

# Computer Vision i OCR
opencv-python==4.8.1.78
pytesseract==0.3.10
Pillow==10.1.0
numpy==1.26.2

# YOLO
ultralytics==8.0.227
torch==2.1.1
torchvision==0.16.1

# Baza danych
supabase==2.0.3
python-dotenv==1.0.0

# Narzędzia pomocnicze
PyYAML==6.0.1
python-multipart==0.0.6
